[TASK] Migrate controller tests to functional tests (deleteAction)
It is considered best practice to cover controllers via functional
instead of unit tests.
We therefore migrate the existing unit tests to functional tests.

The migration is split into multiple batches for smaller changes.
This one covers tests related to `deleteAction()`.

Relates: #1569
