+++ Tests/Functional/Controller/Assertions/Database/FrontEndEditorController/Delete/Deleted.csv
@@ -1,0 +1,5 @@
+"tx_tea_domain_model_tea"
+,"uid","pid","title","owner","deleted"
+,1,2,"Godesberger Burgtee",1,1
+,2,2,"Oolong",2,0
+,3,2,"Sencha",0,0
