+++ Tests/Functional/Controller/FrontEndEditorControllerTest.php
@@ -6,6 +6,7 @@ namespace TTN\Tea\Tests\Functional\Controller;
 
 use PHPUnit\Framework\Attributes\CoversClass;
 use PHPUnit\Framework\Attributes\Test;
+use Psr\Http\Message\ResponseInterface;
 use TTN\Tea\Controller\FrontEndEditorController;
 use TYPO3\TestingFramework\Core\Functional\Framework\Frontend\InternalRequest;
 use TYPO3\TestingFramework\Core\Functional\Framework\Frontend\InternalRequestContext;
@@ -13,20 +14,24 @@ use TYPO3\TestingFramework\Core\Functional\Framework\Frontend\InternalRequestCon
 #[CoversClass(FrontEndEditorController::class)]
 final class FrontEndEditorControllerTest extends AbstractFrontendControllerTestCase
 {
-    private const PAGE_UID = 1;
+    private const UID_OF_PAGE = 1;
+    private const UID_OF_TEA_OWNED_BY_LOGGED_IN_USER = '1';
+    private const UID_OF_TEA_OWNED_BY_FOREIGN_USER = '2';
+    private const UID_OF_TEA_WITHOUT_OWNER = '3';
 
     protected function setUp(): void
     {
         parent::setUp();
 
         $this->importCSVDataSet(__DIR__ . '/Fixtures/Database/FrontEndEditorController/ContentElement.csv');
+        $this->importCSVDataSet(__DIR__ . '/Fixtures/Database/FrontEndEditorController/Teas.csv');
         $this->importCSVDataSet(__DIR__ . '/Fixtures/Database/FrontEndEditorController/FrontendUser.csv');
     }
 
     #[Test]
     public function indexActionForNoLoggedInUserRendersErrorMessage(): void
     {
-        $request = (new InternalRequest())->withPageId(self::PAGE_UID);
+        $request = (new InternalRequest())->withPageId(self::UID_OF_PAGE);
 
         $html = (string)$this->executeFrontendSubRequest($request)->getBody();
 
@@ -36,8 +41,6 @@ final class FrontEndEditorControllerTest extends AbstractFrontendControllerTestC
     #[Test]
     public function indexActionForLoggedInUserRendersTeaOwnedByTheLoggedInUser(): void
     {
-        $this->importCSVDataSet(__DIR__ . '/Fixtures/Database/FrontEndEditorController/Teas.csv');
-
         $html = $this->getHtmlWithLoggedInUser();
 
         self::assertStringContainsString('Godesberger Burgtee', $html);
@@ -46,8 +49,6 @@ final class FrontEndEditorControllerTest extends AbstractFrontendControllerTestC
     #[Test]
     public function indexActionForLoggedInUserDoesNotRenderTeaOwnedByOtherUser(): void
     {
-        $this->importCSVDataSet(__DIR__ . '/Fixtures/Database/FrontEndEditorController/Teas.csv');
-
         $html = $this->getHtmlWithLoggedInUser();
 
         self::assertStringNotContainsString('Oolong', $html);
@@ -56,18 +57,68 @@ final class FrontEndEditorControllerTest extends AbstractFrontendControllerTestC
     #[Test]
     public function indexActionForLoggedInUserDoesNotRenderTeaWithoutOwner(): void
     {
-        $this->importCSVDataSet(__DIR__ . '/Fixtures/Database/FrontEndEditorController/Teas.csv');
-
         $html = $this->getHtmlWithLoggedInUser();
 
         self::assertStringNotContainsString('Sencha', $html);
     }
 
-    private function getHtmlWithLoggedInUser(): string
+    #[Test]
+    public function deleteActionWithOwnTeaRemovesProvidedTea(): void
     {
-        $request = (new InternalRequest())->withPageId(self::PAGE_UID);
+        $this->executeRequestWithLoggedInUser([
+            'tx_tea_teafrontendeditor[action]' => 'delete',
+            'tx_tea_teafrontendeditor[tea][__identity]' => self::UID_OF_TEA_OWNED_BY_LOGGED_IN_USER,
+        ]);
+
+        $this->assertCSVDataSet(__DIR__ . '/Assertions/Database/FrontEndEditorController/Delete/Deleted.csv');
+    }
+
+    #[Test]
+    public function deleteActionWithTeaFromOtherUserThrowsException(): void
+    {
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('You do not have the permissions to edit this tea.');
+        $this->expectExceptionCode(1687363749);
+
+        $this->executeRequestWithLoggedInUser([
+            'tx_tea_teafrontendeditor[action]' => 'delete',
+            'tx_tea_teafrontendeditor[tea][__identity]' => self::UID_OF_TEA_OWNED_BY_FOREIGN_USER,
+        ]);
+    }
+
+    #[Test]
+    public function deleteActionWithTeaWithoutOwnerThrowsException(): void
+    {
+        $this->expectException(\RuntimeException::class);
+        $this->expectExceptionMessage('You do not have the permissions to edit this tea.');
+        $this->expectExceptionCode(1687363749);
+
+        $this->executeRequestWithLoggedInUser([
+            'tx_tea_teafrontendeditor[action]' => 'delete',
+            'tx_tea_teafrontendeditor[tea][__identity]' => self::UID_OF_TEA_WITHOUT_OWNER,
+        ]);
+    }
+
+    /**
+     * @param array<string, string> $queryParameters
+     */
+    private function getHtmlWithLoggedInUser(array $queryParameters = []): string
+    {
+        return (string)$this->executeRequestWithLoggedInUser($queryParameters)->getBody();
+    }
+
+    /**
+     * @param array<string, string> $queryParameters
+     */
+    private function executeRequestWithLoggedInUser(array $queryParameters = []): ResponseInterface
+    {
+        $request = (new InternalRequest())
+            ->withPageId(self::UID_OF_PAGE)
+            ->withQueryParameters($queryParameters)
+        ;
+
         $context = (new InternalRequestContext())->withFrontendUserId(1);
 
-        return (string)$this->executeFrontendSubRequest($request, $context)->getBody();
+        return $this->executeFrontendSubRequest($request, $context);
     }
 }
