+++ Tests/Unit/Controller/FrontEndEditorControllerTest.php
@@ -285,59 +285,4 @@ final class FrontEndEditorControllerTest extends UnitTestCase
 
         $this->subject->updateAction($tea);
     }
-
-    #[Test]
-    public function deleteActionWithOwnTeaRemovesProvidedTea(): void
-    {
-        $userUid = 5;
-        $this->setUidOfLoggedInUser($userUid);
-        $tea = new Tea();
-        $tea->setOwnerUid($userUid);
-        $this->stubRedirect('index');
-
-        $this->teaRepositoryMock->expects(self::once())->method('remove')->with($tea);
-
-        $this->subject->deleteAction($tea);
-    }
-
-    #[Test]
-    public function deleteActionWithOwnTeaRedirectsToIndexAction(): void
-    {
-        $userUid = 5;
-        $this->setUidOfLoggedInUser($userUid);
-        $tea = new Tea();
-        $tea->setOwnerUid($userUid);
-
-        $this->mockRedirect('index');
-
-        $this->subject->deleteAction($tea);
-    }
-
-    #[Test]
-    public function deleteActionWithTeaFromOtherUserThrowsException(): void
-    {
-        $this->setUidOfLoggedInUser(1);
-        $tea = new Tea();
-        $tea->setOwnerUid(2);
-
-        $this->expectException(\RuntimeException::class);
-        $this->expectExceptionMessage('You do not have the permissions to edit this tea.');
-        $this->expectExceptionCode(1687363749);
-
-        $this->subject->deleteAction($tea);
-    }
-
-    #[Test]
-    public function deleteActionWithTeaWithoutOwnerThrowsException(): void
-    {
-        $this->setUidOfLoggedInUser(1);
-        $tea = new Tea();
-        $tea->setOwnerUid(0);
-
-        $this->expectException(\RuntimeException::class);
-        $this->expectExceptionMessage('You do not have the permissions to edit this tea.');
-        $this->expectExceptionCode(1687363749);
-
-        $this->subject->deleteAction($tea);
-    }
 }
